requirements.txt
flask==2.3.2
werkzeug==2.3.6
opencv-python==4.8.0.76
numpy==1.26.4
torch==2.0.1
torchvision==0.15.2
ultralytics==8.3.130
keras==3.5.0
tensorflow==2.18.0
ffmpeg-python==0.2.0
scikit-learn==1.2.2